Fix bugs when tibble input
This is hack to ensure that `[` will work even if the input data is a `tbl_df`. Right now much of the code assumes that `drop = TRUE` is the default for `[`, while `tbl_df` do not drop dimensions when subsetting. This was brought up in #6, but this PR fixes it.